requirements.txt
# network-qualification-suite — Python 3.10+
psutil>=5.9.0
PyYAML>=6.0.1
rich>=13.7.0
